+++ auger_cli/commands/experiments/api.py
@@ -3,7 +3,7 @@ import os
 import click
 
 from ...formatter import print_line, print_record, wait_for_task_result
-from ...utils import request_list, get_uid
+from ...utils import request_list, get_uid, load_dataframe_from_file, save_dict_to_csv
 from ...auger_config import AugerConfig
 from ...constants import REQUEST_LIMIT
 
@@ -28,6 +28,7 @@ from ..experiment_sessions.api import (
     experiment_session_attributes
 )
 from ..clusters.api import read_cluster, cluster_attributes
+from ..pipelines.api import read_pipeline, pipeline_attributes
 from ..predictions.api import (
     read_prediction,
     create_prediction
@@ -125,6 +126,13 @@ def get_or_create_experiment(ctx, project_id):
     experiment = read_experiment(ctx, project_id, experiment_name)
     if experiment.get('id') is not None:
         return experiment
+
+    if len(ctx.config.get_evaluation()) == 0:
+        return {}
+
+    if len(ctx.config.get_evaluation().get('data_path', ''))==0:
+        print_line("To create experiment {}, evaluation_options should contain data_path.".format(experiment_name))
+        return {}
 
     print_line(
         'Experiment {} does not exist. Creating ...'.format(experiment_name))
@@ -258,7 +266,9 @@ def export_model_experiment(ctx, trial_id, deploy=False):
             ],
             poll_interval=10
         )
-
+        
+        pipeline = read_pipeline(ctx, trial_id)        
+        print("Pipeline {} status: {}; error: {}".format(pipeline.get('id'), pipeline.get('status'), pipeline.get('error_message')))
         return trial_id
     else:    
         model_path = create_cluster_task_ex(ctx, project_id,
@@ -275,21 +285,21 @@ def predict_experiment(ctx, pipeline_id, trial_id, file):
     if pipeline_id is None:
         pipeline_id = export_model_experiment(ctx, trial_id, deploy=True)
 
-    import csv
-    header = None
-    data = []
-    with open(file, newline='') as csvfile:
-        reader = csv.reader(csvfile)
-        for row in reader:
-            if header is None:
-                header = row
-            else:
-                data.append(row)
+    pipeline = read_pipeline(ctx, pipeline_id)
+    if pipeline.get('status') != 'ready':
+        print_line("Pipeline is not ready or has issues. Try to create another one.", err = True)
+        print_record(pipeline, pipeline_attributes)
+        return
 
+    df = load_dataframe_from_file(file)
 
-    prediction_id = create_prediction(ctx, pipeline_id, data, header)
+    prediction_id = create_prediction(ctx, pipeline_id, df.values.tolist(), df.columns.get_values().tolist())
     result = read_prediction(ctx, prediction_id)
-    print(result)
+
+    predict_path = os.path.splitext(file)[0] + "_predicted.csv"
+    save_dict_to_csv(result.get('result', {}), predict_path)
+
+    print("Predcition result saved to file: %s"%predict_path)
 
 def monitor_leaderboard_experiment(ctx, name):
     pass