+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Release History
 ===============
 
+# v0.1.3
+- Added 'auger projects open' command to open project url in the browser
+- Support error handling in 'auger experiments predict' command
+
 # v0.1.2
 - Automated release process